# Speed up json serialization with orjson and custom FastAPI responses