--- a/src/dstack/_internal/utils/json_utils.py
+++ b/src/dstack/_internal/utils/json_utils.py
@@ -1,4 +1,3 @@
-import asyncpg.pgproto.pgproto
 import orjson
 from pydantic import BaseModel